moved css files to styles folder
